Handle "Note On" events with velocity 0
A lot of MIDI instruments (almost every keyboard I ever played!) are
using "running status" and emit "Note On" with a velocity of zero
when a key is released (instead of "Note Off").

This causes every note to sound twice (once when pressed, once when
released).

This patch addresses this. Sorry if it's not very Node-ish; feel
free to rewrite, discard, or ignore :)